requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
apify_client==1.8.1
apify_shared==1.1.2
backports.tarfile==1.2.0
bidict==0.23.1
blinker==1.8.2
build==1.2.2.post1
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
distro==1.9.0
docutils==0.21.2
fastapi==0.115.2
firebase-admin==6.5.0
Flask==3.0.3
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.67.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.4
jiter==0.6.1
keyring==25.4.1
lazy_loader==0.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.5.0
msgpack==1.1.0
nh3==0.2.18
openai==1.52.0
packaging==24.1
pillow==11.0.0
pipdeptree==2.16.2
pkginfo==1.10.0
platformdirs==4.3.6
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
PyPDF2==3.0.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.12
python-socketio==5.11.4
readme_renderer==44.0
redis==5.1.1
reflex==0.6.3
reflex-chakra==0.6.2
reflex-hosting-cli==0.1.13
reportlab==4.2.5
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.2
rsa==4.9
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.40.0
starlette-admin==0.14.1
tabulate==0.9.0
tomlkit==0.13.2
tqdm==4.66.5
twine==5.1.1
typer==0.12.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
websockets==13.1
Werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0
zipp==3.20.2
